+++ app/Http/Controllers/Users/Data.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Users;
 
 use App\Http\Controllers\Controller;
 use Enraiged\Users\Models\User;
-use Enraiged\Users\Tables\Builders\UserIndex;
+use Enraiged\Users\Tables\UserIndex;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;