+++ app/Http/Controllers/Users/Show.php
@@ -3,9 +3,11 @@
 namespace App\Http\Controllers\Users;
 
 use App\Http\Controllers\Controller;
+use Enraiged\Users\Enums\Roles;
 use Enraiged\Users\Responses\UserInertiaResponse;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Http\Request;
+use Inertia\Response as InertiaResponse;
 
 class Show extends Controller
 {
@@ -15,13 +17,16 @@ class Show extends Controller
      *  @param  \Illuminate\Http\Request  $request
      *  @return \Inertia\Response
      */
-    public function __invoke(Request $request)//: InertiaResponse
+    public function __invoke(Request $request): InertiaResponse
     {
         $model = config('auth.providers.users.model');
+        $roles = config('auth.providers.roles.enum', Roles::class);
 
         $user = $request->routeIs('my.*')
             ? $request->user()
-            : $model::findOrFail($request->user);
+            : ($request->user()->role->is($roles::find('Administrator'))
+                ? $model::withTrashed()->findOrFail($request->user)
+                : $model::findOrFail($request->user));
 
         $this->authorize('show', $user);