+++ resources/js/components/tables/VueTable.vue
@@ -124,6 +124,7 @@
                 :field="name"
                 :frozen="columnFrozen(column)"
                 :header="column.label"
+                :hidden="column.hidden"
                 :key="name"
                 :sortable="column.sortable && column.sortable !== false">
                 <template #body="column" v-if="column.custom">
@@ -162,6 +163,7 @@
                 </primevue-row>
             </primevue-column-group>
         </primevue-datatable>
+        <primevue-confirm-dialog></primevue-confirm-dialog>
     </div>
 </template>
 
@@ -169,10 +171,10 @@
 import { error } from '@/handlers/errors';
 import { router } from '@inertiajs/vue3';
 import { trans as i18n } from 'laravel-vue-i18n';
-import { useMessages } from '@/handlers/messages';
 import PrimevueButton from 'primevue/button';
 import PrimevueColumn from 'primevue/column';
 import PrimevueColumnGroup from 'primevue/columngroup';
+import PrimevueConfirmDialog from 'primevue/confirmdialog';
 import PrimevueDatatable from 'primevue/datatable';
 import PrimevueInputgroup from 'primevue/inputgroup';
 import PrimevueInputgroupAddon from 'primevue/inputgroupaddon';
@@ -182,13 +184,12 @@ import PrimevueSelect from 'primevue/select';
 import PrimevueTooltip from 'primevue/tooltip';
 import VueTableFilters from './VueTableFilters.vue';
 
-const { flash, flashSuccess } = useMessages();
-
 export default {
     components: {
         PrimevueButton,
         PrimevueColumn,
         PrimevueColumnGroup,
+        PrimevueConfirmDialog,
         PrimevueDatatable,
         PrimevueInputgroup,
         PrimevueInputgroupAddon,
@@ -204,6 +205,8 @@ export default {
 
     inject: [
         'app',
+        //'flash',
+        //'flashSuccess',
         'route',
     ],
 
@@ -273,9 +276,6 @@ export default {
         sums: [],
         timer: null,
     }),
-
-    setup (props, { emit }) {
-    },
 
     computed: {
         batchActions() {
@@ -481,14 +481,14 @@ export default {
                 .then((response) => {
                     const { data, status } = response;
                     if (data.success) {
-                        flashSuccess(data.success);
+                        // this.flashSuccess(data.success);
                         this.fetch();
                         if (typeof success === 'function') {
                             success();
                         }
                     } else if (data.warn) {
-                        const severity = Object.keys(data)[0];
-                        flash({ severity, content: data.warn, expiry: 5000 });
+                        // const severity = Object.keys(data)[0];
+                        // this.flash({ severity, content: data.warn, expiry: 5000 });
                     }
                     if (status === 205) {
                         router.get('/');
@@ -563,7 +563,10 @@ export default {
         defaultFilters() {
             if (this.template.filters) {
                 Object.keys(this.template.filters).forEach((filter) => {
-                    this.filters[filter] = this.template.filters[filter].value || null;
+                    const template = this.template.filters[filter];
+                    console.log(template);
+                    this.filters[filter] = template.value
+                        || (template.type === 'select' && template.multiple ? [] : null);
                 });
             }
         },
@@ -691,7 +694,7 @@ export default {
 
         rows(event) {
             this.pagination.rows = event.value;
-            this.fetch()
+            this.fetch();
         },
 
         sort(event) {